+++ src/Options.js
@@ -81,7 +81,7 @@ const Options = {
 	
 	/**
 	 * A NodeJS module descriptor to a file or module that exports an object to use as an
-	 * adapter for a CSS transpiler such as LESS or SASS.
+	 * adaptor for a CSS transpiler such as LESS or SASS.
 	 *
 	 * Relative to `base` if beginning with `./` or `../`
 	 *
@@ -92,7 +92,7 @@ const Options = {
 	 * Property      | Type     | Description
 	 * --------------|----------|--------------
 	 * `sourcePaths` | string[] | An array of source file paths, relative to the parent directory of `entryPath`.
-	 * `entryPath`   | string   | The absolute path to the entry file for the adapter to write, and for the transpiler itself to read.
+	 * `entryPath`   | string   | The absolute path to the entry file for the adaptor to write, and for the transpiler itself to read.
 	 * `outputPath`  | string   | The absolute path to the css file for the transpiler to write.
 	 * `isDev`       | boolean  | Whether to do a development build, presumably with a sourcemap.
 	 *