+++ test/test-buildTranspilerFile.js
@@ -10,7 +10,7 @@ const buildTranspilerFile = require('../src/buildTranspilerFile.js');
 
 describe(name, async () => {
 
-	it('should send paths to adapter function and return asset paths', async () => {
+	it('should send paths to adaptor function and return asset paths', async () => {
 		await cloneRun(async (base, box) => { // eslint-disable-line no-unused-vars
 			let before = box.snapshot();