+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - mojl-terser as optional dependency
+- mojl-clean-css as optional dependency
 
 ### Changed
 - `jsMinifierAdaptor` is now `mojl-terser` by default